requirements.txt
aniso8601==7.0.0
atomicwrites==1.3.0
attrs==19.1.0
Click==7.0
coverage==4.5.1
coveralls==1.8.2
Flask==1.1.1
flask-restplus==0.12.1
importlib-metadata==0.19
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.2
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.1
pep8==1.7.1
pluggy==0.12.0
psycopg2-binary==2.8.3
py==1.8.0
PyJWT==1.7.1
pyparsing==2.4.2
pyrsistent==0.15.4
python-dotenv==0.10.3
pytz==2019.2
six==1.12.0
wcwidth==0.1.7
Werkzeug==0.15.5
zipp==0.5.2
pytest==4.2.0
pytest-cov==2.6.1
strict-rfc3339==0.7
